requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
xgboost==2.1.1
imbalanced-learn==0.12.3
google-cloud-bigquery==3.25.0
google-cloud-bigquery-storage==2.25.0
db-dtypes==1.3.0
anthropic==0.34.2
langchain
langchain-anthropic
langchain-community
langgraph
fastapi==0.115.0
uvicorn==0.30.6
streamlit==1.38.0
python-dotenv==1.0.1
pydantic==2.9.2
loguru==0.7.2
jupyter==1.1.1

